added some refactoring
