+++ .gitignore
@@ -4,6 +4,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
+logs/
 
 # Compiled binary addons (https://nodejs.org/api/addons.html)
 build/Release
