+++ server/errors/http404NotFound.error.js
@@ -1,4 +1,4 @@
-const StatusCodes = require('../utils/httpStatusCode');
+const {StatusCodes} = require('../utils');
 const CustomAPIError = require('./customApi.error');
 
 class HTTP404NotFoundError extends CustomAPIError {
