+++ server/middlewares/errorHandler.middleware.js
@@ -1,18 +1,33 @@
-const StatusCodes = require('../utils/httpStatusCode');
+const logger = require('../logger');
+const {StatusCodes} = require('../utils');
 
 const errorHandlerMiddleware = (err, req, res, next) => {
   let customError = {
     statusCode: err.statusCode || StatusCodes.INTERNAL_SERVER_ERROR,
     msg: err.message || 'Something went wrong! Please try again later.',
   };
+  console.log(err.name);
+  switch (err.name){
+    case 'SequelizeValidationError':
+      customError.msg = "Database Validation Error!";
+      customError.statusCode = StatusCodes.BAD_REQUEST;
+      break;
+    case 'SequelizeDatabaseError':
+      customError.msg = "Sequelize Database  Error!";
+      customError.statusCode = StatusCodes.BAD_REQUEST;
+      break;
+    case 'SequelizeUniqueConstraintError':
+      customError.statusCode = StatusCodes.BAD_REQUEST;
+      customError.msg = err.errors[0].message;
+      break;
+    default:
 
-  if (err.name === 'SequelizeValidationError' || err.name == 'SequelizeDatabaseError')  customError.statusCode = 400;
-  if (err.name === 'SequelizeUniqueConstraintError') {
-    customError.msg = err.errors[0].message;
-    customError.statusCode = 400;
   }
-
-  return res.status(customError.statusCode).json({ message: customError.msg });
+  logger.error(
+    'Error message from the centralized error-handling component',
+    err,
+  );
+  return res.status(customError.statusCode).send({ message: customError.msg });
 };
 
 module.exports = errorHandlerMiddleware;
