+++ server/utils/index.js
@@ -1,10 +1,12 @@
 const {createJWT} = require('./jwt');
 const {generateHashPassword, comparePassword} = require('./hashing')
 const validator = require('./validator')
+const StatusCodes = require('./httpStatusCode')
 
 module.exports = {
     createJWT,
     generateHashPassword,
     comparePassword,
-    validator
+    validator,
+    StatusCodes
 }
